move bin to where they should be
