+++ crates/maa-server/Cargo.toml
@@ -8,11 +8,9 @@ repository.workspace = true
 
 [[bin]]
 name = "server"
-path = "src/server.rs"
 
 [[bin]]
 name = "client"
-path = "src/client.rs"
 
 [features]
 default = ["client","unix-socket"]
