+++ crates/maa-server/src/bin/client.rs
